fix: moved Code from RegisterPage to Register Component
