+++ client/src/services/auth.tsx
@@ -26,7 +26,7 @@ export const login = async (email: string, password: string) => {
   }
 };
 
-export const register = async (
+export const registerPOST = async (
   username: string,
   email: string,
   password: string
